Fixed broken test cases (adding new DateFactoryTest and moving existing test cases from DateTimeFactoryTest)
